Fix fs async issue with tests
