+++ test/atRule.spec.js
@@ -1,11 +1,10 @@
 const compareCritical = require('./compareCritical');
-const { fsTimeout, normalizeOpts } = require('./utils');
+const { normalizeOpts } = require('./utils');
 const preTest = require('./preTest');
 
 const opts = normalizeOpts();
 beforeAll(async () => {
-  preTest('atRule', opts);
-  await fsTimeout();
+  await preTest('atRule', opts);
 });
 
 describe('tests for @critical rule', () => {
