+++ test/scope.spec.js
@@ -1,11 +1,10 @@
 const compareCritical = require('./compareCritical');
-const { fsTimeout, normalizeOpts } = require('./utils');
+const { normalizeOpts } = require('./utils');
 const preTest = require('./preTest');
 
 const opts = normalizeOpts();
 beforeAll(async () => {
-  preTest('scope', opts);
-  await fsTimeout();
+  await preTest('scope', opts);
 });
 
 describe('tests for `critical-selector: scope`', () => {
